remove interactor gem, refactor code
